PB-271: Put disclaimer in separate component
