+++ src/modules/i18n/locales/de.json
@@ -173,7 +173,7 @@
     "emapis_service_link_href": "mailto:info@blw.admin.ch",
     "emapis_service_link_label": "info@blw.admin.ch",
     "embed_map": "Karte einbetten",
-    "empty_description" : "Leere Beschreibung",
+    "empty_description" : "TODO",
     "energie": "Energie",
     "energie_service_link_href": "http://www.bfe.admin.ch/geoinformation/index.html?lang=de",
     "energie_service_link_label": "www.map.energie.admin.ch",
