+++ src/modules/i18n/locales/fr.json
@@ -173,7 +173,7 @@
     "emapis_service_link_href": "mailto:info@blw.admin.ch",
     "emapis_service_link_label": "info@blw.admin.ch",
     "embed_map": "Intégrer la carte",
-    "empty_description" : "Description vide",
+    "empty_description" : "TODO",
     "energie": "Énergie",
     "energie_service_link_href": "http://www.bfe.admin.ch/geoinformation/index.html?lang=fr",
     "energie_service_link_label": "www.map.energie.admin.ch",
