+++ src/modules/i18n/locales/it.json
@@ -173,7 +173,7 @@
     "emapis_service_link_href": "mailto:info@blw.admin.ch",
     "emapis_service_link_label": "info@blw.admin.ch",
     "embed_map": "Incorpora mappa",
-    "empty_description" : "descrizione vuota",
+    "empty_description" : "TODO",
     "energie": "Energia",
     "energie_service_link_href": "http://www.bfe.admin.ch/geoinformation/index.html?lang=it",
     "energie_service_link_label": "www.map.energia.admin.ch",
